+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.32...@kaizen/draft-modal@5.4.33) (2021-05-13)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.31...@kaizen/draft-modal@5.4.32) (2021-05-11)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [5.4.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.30...@kaizen/draft-modal@5.4.31) (2021-05-05)
 
 **Note:** Version bump only for package @kaizen/draft-modal