+++ draft-packages/tabs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tabs@4.1.35...@kaizen/draft-tabs@4.1.36) (2021-05-13)
+
+**Note:** Version bump only for package @kaizen/draft-tabs
+
+
+
+
+
 ## [4.1.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tabs@4.1.34...@kaizen/draft-tabs@4.1.35) (2021-05-05)
 
 **Note:** Version bump only for package @kaizen/draft-tabs